fix: merge lifecycle helpers into read test file for pkg file-count gate
